# Significant improvement in deserialization speed