--- a/src/main/kotlin/com/fasterxml/jackson/module/kotlin/KotlinFeature.kt
+++ b/src/main/kotlin/com/fasterxml/jackson/module/kotlin/KotlinFeature.kt
@@ -6,7 +6,7 @@ import kotlin.math.pow
 /**
  * @see KotlinModule.Builder
  */
-enum class KotlinFeature(private val enabledByDefault: Boolean) {
+enum class KotlinFeature(internal val enabledByDefault: Boolean) {
     /**
      * This feature represents whether to deserialize `null` values for collection properties as empty collections.
      */
@@ -42,7 +42,7 @@ enum class KotlinFeature(private val enabledByDefault: Boolean) {
      * may contain null values after deserialization.
      * Enabling it protects against this but has significant performance impact.
      */
-    StrictNullChecks(enabledByDefault = false);
+    StrictNullChecks(enabledByDefault = false),
 
     ExperimentalDeserializationBackend(enabledByDefault = false);