--- a/src/main/kotlin/com/fasterxml/jackson/module/kotlin/KotlinValueInstantiator.kt
+++ b/src/main/kotlin/com/fasterxml/jackson/module/kotlin/KotlinValueInstantiator.kt
@@ -29,7 +29,8 @@ internal class KotlinValueInstantiator(
     private val nullToEmptyCollection: Boolean,
     private val nullToEmptyMap: Boolean,
     private val nullIsSameAsDefault: Boolean,
-    private val strictNullChecks: Boolean
+    private val strictNullChecks: Boolean,
+    private val experimentalDeserializationBackend: Boolean
 ) : StdValueInstantiator(src) {
     @Suppress("UNCHECKED_CAST")
     override fun createFromObjectWith(
@@ -188,7 +189,8 @@ internal class KotlinInstantiators(
     private val nullToEmptyCollection: Boolean,
     private val nullToEmptyMap: Boolean,
     private val nullIsSameAsDefault: Boolean,
-    private val strictNullChecks: Boolean
+    private val strictNullChecks: Boolean,
+    private val experimentalDeserializationBackend: Boolean
 ) : ValueInstantiators {
     override fun findValueInstantiator(
         deserConfig: DeserializationConfig,
@@ -197,7 +199,15 @@ internal class KotlinInstantiators(
     ): ValueInstantiator {
         return if (beanDescriptor.beanClass.isKotlinClass()) {
             if (defaultInstantiator is StdValueInstantiator) {
-                KotlinValueInstantiator(defaultInstantiator, cache, nullToEmptyCollection, nullToEmptyMap, nullIsSameAsDefault, strictNullChecks)
+                KotlinValueInstantiator(
+                    defaultInstantiator,
+                    cache,
+                    nullToEmptyCollection,
+                    nullToEmptyMap,
+                    nullIsSameAsDefault,
+                    strictNullChecks,
+                    experimentalDeserializationBackend
+                )
             } else {
                 // TODO: return defaultInstantiator and let default method parameters and nullability go unused?  or die with exception:
                 throw IllegalStateException("KotlinValueInstantiator requires that the default ValueInstantiator is StdValueInstantiator")