--- a/src/test/kotlin/com/fasterxml/jackson/module/kotlin/DslTest.kt
+++ b/src/test/kotlin/com/fasterxml/jackson/module/kotlin/DslTest.kt
@@ -2,11 +2,8 @@ package com.fasterxml.jackson.module.kotlin
 
 import com.fasterxml.jackson.core.json.JsonReadFeature
 import com.fasterxml.jackson.core.json.JsonWriteFeature
-import com.fasterxml.jackson.module.kotlin.KotlinFeature.NullIsSameAsDefault
-import com.fasterxml.jackson.module.kotlin.KotlinFeature.NullToEmptyCollection
-import com.fasterxml.jackson.module.kotlin.KotlinFeature.NullToEmptyMap
+import com.fasterxml.jackson.module.kotlin.KotlinFeature.*
 import com.fasterxml.jackson.module.kotlin.KotlinFeature.SingletonSupport
-import com.fasterxml.jackson.module.kotlin.KotlinFeature.StrictNullChecks
 import com.fasterxml.jackson.module.kotlin.SingletonSupport.CANONICALIZE
 import org.junit.Assert.assertNotNull
 import org.junit.Test
@@ -37,6 +34,7 @@ class DslTest {
             enable(NullIsSameAsDefault)
             enable(SingletonSupport)
             enable(StrictNullChecks)
+            enable(ExperimentalDeserializationBackend)
         }
 
         assertNotNull(module)
@@ -46,6 +44,7 @@ class DslTest {
         assertTrue(module.nullIsSameAsDefault)
         assertEquals(module.singletonSupport, CANONICALIZE)
         assertTrue(module.strictNullChecks)
+        assertTrue(module.experimentalDeserializationBackend)
     }
 
     @Test