… pom.xml
… src/main/kotlin/com/fasterxml/jackson/module/kotlin/ReflectionCache.kt