@k163377 2021-05-03
@cowtowncoder 
Thanks for the reply.

> One thing that would be nice, but I am not sure if it is possible, would be to have choice of two backends for 2.13: "old" one for backwards compatibility, and "new" one with optimizations. Users would have to opt-in (enable) use of the new, faster version, and only use it if it works for them. If and when issues are resolved, in 2.14 and later new backend would become the default.

That is certainly true.
I too think that this change is so large that it would be preferable to be able to switch between the two.

I tried to add a property to switch between the two.
How about it?
2bc2743
75a1c3e
ace6cd0

> One quick comment: change would definitely need to be against 2.13 branch, to minimize risk of breakage (which I think there is just due to scope).

Understood.
However, since I've already submitted the PR, I'm thinking of having it reviewed as is and changing the branch later, what do you think?

Also, which is a better way to do this change process, creating a new PR or doing a `rebase`?