@k163377 2021-05-05
After creating this PR, I noticed that the reflection process on read is also cacheable.
I redid a [simple benchmark](https://github.com/k163377/jackson-kotlin-benchmark-sample/tree/22c72607a055135b6fff021c1b1eace536559fc6) (with more benchmarks, but little change in the score) using [a branch that incorporates these results](https://github.com/k163377/jackson-module-kotlin/tree/speed_up/more_speed_up), and the results are as follows.

```
Benchmark                                           Mode  Cnt        Score       Error  Units
c.w.constructor.Benchmarks.useDefaultKotlinMapper  thrpt    9  2513352.950 ± 47859.737  ops/s
c.w.constructor.Benchmarks.useJavaMapper           thrpt    9  3054154.327 ± 27503.185  ops/s
c.w.constructor.Benchmarks.useKotlinMapper         thrpt    9  2353756.705 ±  9910.363  ops/s
c.w.factory.Benchmarks.useDefaultKotlinMapper      thrpt    9  2438265.573 ± 57048.905  ops/s
c.w.factory.Benchmarks.useJavaMapper               thrpt    9  3070987.061 ± 32149.819  ops/s
c.w.factory.Benchmarks.useKotlinMapper             thrpt    9  2326640.414 ± 62539.462  ops/s
```

The score is increased to about 4/5 compared to `Java` deserialization.
The results also show that the performance of `StrictNullCheck` is expected to improve significantly as well.

Initially, I intended to make the readout process common between the experimental backend and the conventional backend, but I reverted the commits for commonality in case I want to incorporate this content in the future.
455eae373475ca03989c40b0aef0befff8b49493