@dinomite 2021-05-07
Looking through this in detail now, as you mention it's big so it'll probably take me a few days. All looks good so far, thank you for the clear explanations.

To your last question, yes, a separate PR for the read side makes sense to me.