@dinomite 2021-05-07
@k163377 Would you mind adding some class comments to explain the major intents of the `ArgumentBucket`? It's a bit dense and I'm having trouble parsing what it's intent is.