@k163377 2021-05-08
@dinomite 
I've added `Javadoc` and comments for `ArgumentBucket`.
For other classes, I'll add some comments if the explanation seems insufficient.