@dinomite 2021-05-08
@k163377 Great, thanks for clarifying that bitmasking—I'm a bit dense at understanding such things.  Things look good to me overall, I'll give another look next week.