@cowtowncoder 2021-05-09
One comment/thought inspired by changes here, @dinomite and @k163377 -- now that there are quite a few boolean-valued configuration settings, I wonder if these should grouped as something like `KotlinMapper.Feature`? So that we could pass a set of configuration settings as one `int` internally and (optionally) expose this to users as well.
Alternatively could only use internally at first, keep separate config settings.

Such change should be separate from this PR of course, but I'll introduce idea now. Will create separate Github issue now.