@k163377 2021-05-09
@dinomite 
This change will interfere with the caching of reflection process on read later, so I'd like you to revert it.
074d87d7d52a11f24acb6cc0f1b60af98a319872

https://github.com/FasterXML/jackson-module-kotlin/pull/439#issuecomment-832444648