@dinomite 2021-05-10
Ahh, unfortunate. I was hoping to make those big methods a bit simpler.