@k163377 2021-05-16
I have a branch that has been modified for 2.13 to make it an official PR.
It is ready for `force push` or creation of a new PR.

- Branch: https://github.com/k163377/jackson-module-kotlin/tree/speed_up/pr_2.13
- Diff: https://github.com/FasterXML/jackson-module-kotlin/compare/2.13...k163377:speed_up/pr_2.13