@k163377 2021-06-06
The branch for 2.13 has been re-pushed to incorporate the changes in #454.
https://github.com/FasterXML/jackson-module-kotlin/pull/439#issuecomment-841773663