@dinomite 2021-06-07
@k163377 I think you can go ahead and force-push to this branch and get rid of the separate 2.13 branch. We're not expecting to do any further 2.12 releases ([2.13 is on the horizon](https://github.com/FasterXML/jackson/issues/83))