@k163377 2022-01-03
@dinomite 
I'm going to do the following three things.

- Improve efficiency of argument management and speed up the case of calling factory functions without default arguments.
- Refactoring of the `KVI` argument reading process.
- Fix minor omissions in #512.

For various reasons, the PR on this will probably be issued sometime this weekend or next Monday.