@k163377 2023-01-14
I have released a project that incorporates these improvements.
If you are interested, please give it a try.
https://github.com/ProjectMapK/jackson-module-kogera