@k163377 2025-04-12
It seems to me that function calls via `kotlin-reflect` are now essential to keep the code maintainable while supporting `value class` and other `Kotlin` specific features.
This `PR` is closed, please use `kogera` if you really need it.